refactor: 스타일과 컴포넌트 분리
