+++ package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@emotion/css": "^11.13.5",
     "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
