+++ src/components/ErrorPopup/ErrorPopup.styles.tsx
@@ -1,0 +1,12 @@
+import styled from "@emotion/styled";
+
+export const ErrorPopupWrapper = styled.div`
+  position: absolute;
+  top: 64px;
+  left: 0;
+  width: 100%;
+  text-align: center;
+  color: #000;
+  background-color: #ffc9c9;
+  padding: 12px 0;
+`;
