+++ src/components/Text/Text.styles.tsx
@@ -1,0 +1,25 @@
+import styled from "@emotion/styled";
+
+export const Text = styled.span<{ variant: string; color: string }>`
+  font-style: normal;
+  line-height: normal;
+  color: ${({ color }) => color};
+  ${({ variant }) => {
+    switch (variant) {
+      case "title-1":
+        return `font-size: 24px; font-weight: 700;`;
+      case "title-2":
+        return `font-size: 14px; font-weight: 700;`;
+      case "title-3":
+        return `font-size: 12px; font-weight: 700;`;
+      case "body-1":
+        return `font-size: 14px; font-weight: 500;`;
+      case "body-2":
+        return `font-size: 12px; font-weight: 500;`;
+      case "body-3":
+        return `font-size: 10px; font-weight: 400;`;
+      default:
+        return "";
+    }
+  }}
+`;
