+++ src/components/icons/index.ts
@@ -1,4 +1,4 @@
-export { default as AddCart } from './AddCart';
-export { default as RemoveCart } from './RemoveCart';
-export { default as ShoppingBag } from './ShoppingBag';
-export { default as TopArrow } from './TopArrow';
+export { default as AddCart } from "./AddCart";
+export { default as RemoveCart } from "./RemoveCart";
+export { default as ShoppingBag } from "./ShoppingBag";
+export { default as TopArrow } from "./TopArrow";
