+++ src/constants/images.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_IMAGE_URL = "https://www.next-t.co.kr/public/uploads/7b7f7e2138e29e598cd0cdf2c85ea08d.jpg";
