+++ src/hooks/index.ts
@@ -1,2 +1,2 @@
 export { default as useCartItems } from "./useCartItems";
-export { default as useProducts } from "./useProducts";
+export { default as useProducts } from "./useFetchProducts";
