+++ src/types/index.ts
@@ -1,2 +1,1 @@
-export * from "./cartItem";
-export * from "./product";
+export * from "./response";
